third_party/WebKit/LayoutTests/bluetooth/service/getCharacteristics/characteristics-found.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
  return setBluetoothFakeAdapter('HeartRateAdapter')
    .then(() => requestDeviceWithKeyDown({
      filters: [{services: ['heart_rate']}]}))
    .then(device => device.gatt.connect())
    .then(gattServer => gattServer.getPrimaryService('heart_rate'))
    .then(service => service.getCharacteristics())
    .then(characteristics => {
      assert_equals(characteristics.length, 3);
      assert_equals(characteristics[0].uuid,
                    BluetoothUUID.getCharacteristic('heart_rate_measurement'));
      assert_equals(characteristics[1].uuid,
                    BluetoothUUID.getCharacteristic('body_sensor_location'));
      assert_equals(characteristics[2].uuid,
                    BluetoothUUID.getCharacteristic('body_sensor_location'));
    });
}, 'Find all characteristics in a service.');
</script>
